+++ web-template/src/main/java/org/ehrbase/openehr/sdk/webtemplate/model/WebTemplateInput.java
@@ -23,14 +23,14 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
+import org.ehrbase.openehr.sdk.webtemplate.util.WebTemplateUtils;
 
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class WebTemplateInput implements Serializable {
 
     private String suffix;
     private String type;
-    private final List<WebTemplateInputValue> list = new ArrayList<>();
+    private final List<WebTemplateInputValue> list;
     private Boolean listOpen;
     private WebTemplateValidation validation;
     private String terminology;
@@ -38,7 +38,9 @@ public class WebTemplateInput implements Serializable {
     @JsonSerialize(using = StringToNumberSerializer.class)
     private String defaultValue;
 
-    public WebTemplateInput() {}
+    public WebTemplateInput() {
+        list = new ArrayList<>();
+    }
 
     public WebTemplateInput(WebTemplateInput other) {
         this.suffix = other.suffix;
@@ -51,7 +53,7 @@ public class WebTemplateInput implements Serializable {
         }
         this.terminology = other.terminology;
         this.defaultValue = other.defaultValue;
-        this.list.addAll(other.list.stream().map(WebTemplateInputValue::new).collect(Collectors.toList()));
+        this.list = WebTemplateUtils.cloneList(other.list, WebTemplateInputValue::new);
     }
 
     public String getSuffix() {