+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Note: version releases in the 0.x.y range may introduce breaking changes.
 
+## [2.20.0]
+ ### Changed
+- Improved Walker and validation performance
+- Bumped libraries
+ ### Added 
+ ### Fixed 
+- ACTION.instruction_details now included in web template
+
 ## [2.19.0]
  ### Added 
 - Option to validate compositions for child nodes not defined in template ([631](https://github.com/ehrbase/openEHR_SDK/pull/631))
@@ -454,3 +462,4 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 [2.17.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.16.0...v2.17.0
 [2.18.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.17.0...v2.18.0
 [2.19.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.18.0...v2.19.0
+[2.20.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.19.0...v2.20.0